1.Fix the build error caused by the modification for compiler 6. 2.Revise startup_NUC122.s and retarget.c files to fix build issue. 3.Modify settings of Keil project fo speed up compilation. 4.Revise copyright format for gen. chm file. 5.Fix SPI_Loopback GCC sample code issue.

+++ SampleCode/RegBased/PS2/main.c
@@ -5,8 +5,9 @@
  * $Date: 15/07/24 4:44p $
  * @brief    Show how to control PS/2 mouse movement on the screen.
  * @note
- * Copyright (C) 2014~2015 Nuvoton Technology Corp. All rights reserved.
+ * @copyright SPDX-License-Identifier: Apache-2.0
  *
+ * @copyright Copyright (C) 2014~2015 Nuvoton Technology Corp. All rights reserved.
  ******************************************************************************/
 #include <stdio.h>
 #include "NUC122.h"
@@ -208,7 +209,7 @@ void PS2_IRQHandler(void)
 
                     g_cmd[0] = 0;
 
-                    printf("(PS2->STATUS).TXEMPTY is (%0x)\n", ((PS2_GET_STATUS() & PS2_PS2STATUS_TXEMPTY_Msk) >> PS2_PS2STATUS_TXEMPTY_Pos));
+                    printf("(PS2->STATUS).TXEMPTY is (%0x)\n", (unsigned int)((PS2_GET_STATUS() & PS2_PS2STATUS_TXEMPTY_Msk) >> PS2_PS2STATUS_TXEMPTY_Pos));
 
                     /* Wait Tx ready */
                     if(PS2_GET_STATUS() & PS2_PS2STATUS_TXEMPTY_Msk)
